+++ src/org/ods/orchestration/usecase/JiraUseCase.groovy
@@ -387,6 +387,20 @@ class JiraUseCase {
         this.jira.updateBuildNumber(projectKey, changeId, buildNumber)
     }
 
+    List buildComponentsDataForRelease() {
+        def components = []
+        this.project.data.metadata.repositories.each { repo ->
+            if (repo.include == true) {
+                components.add([
+                    id: repo.id,
+                    commit: repo.data?.git?.commit,
+                    failed: repo.data?.failedStage ? true : false,
+                ])
+            }
+        }
+        return components
+    }
+
     void updateJiraReleaseStatusResult(String message, boolean isError) {
         if (!this.jira) {
             logger.warn("updateJiraReleaseStatusResult: Could *NOT* update release status result because jira has invalid value.")
@@ -418,6 +432,7 @@ class JiraUseCase {
             ],
             status: status,
             env: env,
+            components: this.buildComponentsDataForRelease(),
             startDateTimestamp: this.steps.currentBuild.startTimeInMillis,
         ]