+++ src/org/ods/orchestration/usecase/JiraUseCase.groovy
@@ -391,10 +391,11 @@ class JiraUseCase {
         def components = []
         this.project.data.metadata.repositories.each { repo ->
             if (repo.include == true) {
+                def gitSHA = repo.data?.git?.commit
                 components.add([
                     id: repo.id,
-                    commit: repo.data?.git?.commit,
-                    failed: repo.data?.failedStage ? true : false,
+                    commit: gitSHA ? gitSHA : 'N/A',
+                    status: gitSHA ? (repo.data?.failedStage ? 'FAILURE' : 'SUCCESS') : 'N/A',
                 ])
             }
         }